Remove parts that are purely story
Doing this serves two purposes. 1, it reduces the amount of crap I have to scroll through. I've kept the parts that are useful to reference when I'm offline. 2, it somewhat ensures this repository doesn't end up as a replacement for the website, as the descriptions aren't freely licensed and it would certainly not be fair dealing/use to copy them in their entirety.